requirements.txt
flask>=2.3.3
werkzeug>=2.3.7
flask-cors==3.0.10
python-dotenv>=1.0.0
openai>=1.55.3
requests>=2.32.3
pyyaml==6.0.1
crewai>=0.95.0
instructor>=1.7.2
litellm>=1.57.8
langchain>=0.1.0
langchain-openai>=0.0.2
PyPDF2==3.0.1
faiss-cpu==1.9.0.post1
numpy>=1.24.0